Break apart ros_subscriber test translation unit (#212)
