+++ ros_ign_bridge/test/subscribers/ros_subscriber/nav_msgs_subscriber.cpp
@@ -1,0 +1,34 @@
+// Copyright 2021 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+#include <gtest/gtest.h>
+
+#include <chrono>
+
+#include "ros_subscriber.hpp"
+
+using ros_subscriber::MyTestClass;
+
+/////////////////////////////////////////////////
+TEST(ROSSubscriberTest, Odometry)
+{
+  MyTestClass<nav_msgs::msg::Odometry> client("odometry");
+
+  using namespace std::chrono_literals;
+  ros_ign_bridge::testing::waitUntilBoolVarAndSpin(
+    ros_subscriber::TestNode(), client.callbackExecuted, 10ms, 200);
+
+  EXPECT_TRUE(client.callbackExecuted);
+}
